backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1
google-generativeai==0.8.4
numpy==2.1.3
gmsh==4.14.1
pymeshfix==0.18.0
trimesh==4.8.3
